Backend Physics calculations correct

+++ Prototypes/OrbitSim/src/sim/simobject/Sun.java
@@ -24,7 +24,7 @@ public class Sun extends ObjectInSpace {
 	 * @param color
 	 * @param brightness
 	 */
-	public Sun(long mass, long density, Point3D position, Color color, long brightness) {
+	public Sun(double mass, long density, Point3D position, Color color, long brightness) {
 		super(mass, 0, density, 0, position, color, null);
 		this.brightness = brightness;
 	}
